@@ -1,4 +1,4 @@
-name: infomahasiswa
+name: belajarbersama
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.7.0
+  sdk: ^3.8.0-278.1.beta
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,11 +30,20 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  firebase_core: ^3.11.0
+  firebase_database: ^11.3.6
+  firebase_auth: ^5.5.4
   cupertino_icons: ^1.0.8
+  google_sign_in: ^6.2.1
+  fl_chart: ^1.0.0
+  geolocator: ^14.0.1
+  image_picker: ^1.0.4
+  firebase_storage: 12.4.6
+  cloud_firestore: 5.6.8
+  cloudinary_flutter: ^1.0.0
+  cloudinary_url_gen: ^1.0.0
+  http: ^1.2.1
+  flutter_image_compress: ^2.2.0
 
 dev_dependencies:
   flutter_test:
@@ -45,7 +54,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -53,16 +62,17 @@ dev_dependencies:
 # The following section is specific to Flutter packages.
 flutter:
 
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
+  fonts:
+    - family: SFUIDisplay
+      fonts:
+        - asset: lib/assets/SFUIDisplay-Regular.otf
+        - asset: lib/assets/SFUIDisplay-Bold.ttf
+
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/profil.jpg
-    - assets/ui3.png
-  #   - images/a_dot_ham.jpeg
+    - lib/assets/landing2.png
+    - lib/assets/google2.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
@@ -75,12 +85,12 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  fonts:
-    - family: SF Pro Display
-      fonts:
-        - asset: assets/fonts/SFUIDisplay-Regular.otf
-        - asset: assets/fonts/SFUIDisplay-Bold.ttf
-          weight: 700
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
