@@ -1,35 +1,41 @@
 import 'package:flutter/material.dart';
-import 'screens/home_screen.dart';
-import 'screens/profile_screen.dart';
-import 'screens/courses_screen.dart';
-import 'screens/about_screen.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'firebase_options.dart';
+import 'landing_page.dart';
+import 'main_page.dart';
 
-void main() {
-  // Fungsi utama untuk menjalankan aplikasi
-  runApp(const InfoMahasiswaApp());
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+  runApp(const MyApp());
 }
 
-class InfoMahasiswaApp extends StatelessWidget {
-  const InfoMahasiswaApp({Key? key}) : super(key: key);
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'InfoMahasiswa App',
-      theme: ThemeData(primarySwatch: Colors.blue),
-      // menuju Layar utama aplikasi
-      home: const HomeScreen(),
-      // Daftar rute navigasi aplikasi
-      routes: {
-        '/profile':
-            (context) =>
-                const ProfileScreen(), // Rute ke layar Profil Mahasiswa
-        '/courses':
-            (context) =>
-                const CoursesScreen(), // Rute ke layar Daftar Mata Kuliah
-        '/about':
-            (context) => const AboutScreen(), // Rute ke layar Tentang Aplikasi
-      },
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        fontFamily: 'SFUIDisplay',
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+      ),
+      home: StreamBuilder<User?>(
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return const Scaffold(
+              body: Center(child: CircularProgressIndicator()),
+            );
+          }
+          if (snapshot.hasData) {
+            return const MainPage();
+          }
+          return const LandingPage();
+        },
+      ),
     );
   }
 }
